+++ src/Repositories/AuthenticationRepository.php
@@ -2,7 +2,12 @@
 
 namespace Shellrent\VeeamVspcApiClient\Repositories;
 
-use Shellrent\VeeamVspcApiClient\Payloads\Payload;
+use Shellrent\VeeamVspcApiClient\Payloads\CreateAuthenticationAsymmetricAlgorithmChallengePayload;
+use Shellrent\VeeamVspcApiClient\Payloads\CreateAuthenticationOAuthTokenPayload;
+use Shellrent\VeeamVspcApiClient\Payloads\CreateAuthenticationPkcs12DecryptionPayload;
+use Shellrent\VeeamVspcApiClient\Payloads\CreateAuthenticationRefreshTokenPayload;
+use Shellrent\VeeamVspcApiClient\Payloads\CreateAuthenticationUsernamePasswordPayload;
+use Shellrent\VeeamVspcApiClient\Payloads\EditAuthenticationAsymmetricAlgorithmChallengePayload;
 use Shellrent\VeeamVspcApiClient\Support\CreateGetRequest;
 use Shellrent\VeeamVspcApiClient\Support\CreatePatchRequest;
 use Shellrent\VeeamVspcApiClient\Support\CreatePostRequest;
@@ -21,7 +26,7 @@ class AuthenticationRepository implements Repository {
 	 * Obtain Tokens for Decrypted Challenge Issues access and refresh tokens in response to a decrypted challenge. > Operation is deprecated. We recommend to authorize using the `/token` endpoint.
 	 * Path: /authentication/asymmetricalgorithm
 	 */
-	public function asymmetricAlgorithmCompleteChallenge(?Payload $payload = null, array $query = []): RequestBuilder {
+	public function asymmetricAlgorithmCompleteChallenge(EditAuthenticationAsymmetricAlgorithmChallengePayload $payload, array $query = []): RequestBuilder {
 		$request = $this->createPatchRequest('/authentication/asymmetricalgorithm', $payload);
 		if ($query !== []) {
 			$request->query($query);
@@ -33,7 +38,7 @@ class AuthenticationRepository implements Repository {
 	 * Generate Asymmetric Authentication Challenge Generates a decryption challenge for the specified public key. The challenge must be decrypted in 30 seconds. > Operation is deprecated. We recommend to authorize using the `/users/{userUid}/logins/apikey` or `/token` endpoint.
 	 * Path: /authentication/asymmetricalgorithm
 	 */
-	public function asymmetricAlgorithmChallenge(?Payload $payload = null, array $query = []): RequestBuilder {
+	public function asymmetricAlgorithmChallenge(CreateAuthenticationAsymmetricAlgorithmChallengePayload $payload, array $query = []): RequestBuilder {
 		$request = $this->createPostRequest('/authentication/asymmetricalgorithm', $payload);
 		if ($query !== []) {
 			$request->query($query);
@@ -69,7 +74,7 @@ class AuthenticationRepository implements Repository {
 	 * Decrypts an encrypted PKCS#12 container. > If container does not include private key the server will return an error with the `1210` code.
 	 * Path: /authentication/keys/rsa-pkcs12/decrypt
 	 */
-	public function decryptPkcs12Container(?Payload $payload = null, array $query = []): RequestBuilder {
+	public function decryptPkcs12Container(CreateAuthenticationPkcs12DecryptionPayload $payload, array $query = []): RequestBuilder {
 		$request = $this->createPostRequest('/authentication/keys/rsa-pkcs12/decrypt', $payload);
 		if ($query !== []) {
 			$request->query($query);
@@ -93,7 +98,7 @@ class AuthenticationRepository implements Repository {
 	 * Obtain New Pair of Tokens Returns access and refresh tokens in response to refresh token. > Operation is deprecated. We recommend to authorize using the `/token` endpoint.
 	 * Path: /authentication/refresh
 	 */
-	public function refreshTokenAuthenticate(?Payload $payload = null, array $query = []): RequestBuilder {
+	public function refreshTokenAuthenticate(CreateAuthenticationRefreshTokenPayload $payload, array $query = []): RequestBuilder {
 		$request = $this->createPostRequest('/authentication/refresh', $payload);
 		if ($query !== []) {
 			$request->query($query);
@@ -105,7 +110,7 @@ class AuthenticationRepository implements Repository {
 	 * Authenticate with User Name and Password Issues access and refresh tokens for the specified user name and password. > Operation is deprecated. We recommend to authorize using the `/token` endpoint.
 	 * Path: /authentication/usernamepassword
 	 */
-	public function usernamePasswordAuthenticate(?Payload $payload = null, array $query = []): RequestBuilder {
+	public function usernamePasswordAuthenticate(CreateAuthenticationUsernamePasswordPayload $payload, array $query = []): RequestBuilder {
 		$request = $this->createPostRequest('/authentication/usernamepassword', $payload);
 		if ($query !== []) {
 			$request->query($query);
@@ -117,7 +122,7 @@ class AuthenticationRepository implements Repository {
 	 * OAuth 2.0 Authentication Performs authentication using the OAuth 2.0 Authorization Framework.
 	 * Path: /token
 	 */
-	public function oAuth2IssueToken(?Payload $payload = null, array $query = []): RequestBuilder {
+	public function oAuth2IssueToken(CreateAuthenticationOAuthTokenPayload $payload, array $query = []): RequestBuilder {
 		$request = $this->createPostRequest('/token', $payload);
 		if ($query !== []) {
 			$request->query($query);